rename in_data() method to in()

+++ container_out.cpp
@@ -3,7 +3,7 @@
 namespace encryption_methods {
   void Container::out(std::ofstream& out_file) {
     out_file << "Container contains " << len << " elements." << std::endl;
-    for (size_t i = 0; i < len; ++i) {
+    for (int i = 0; i < len; ++i) {
       out_file << i << ": ";
       array[i]->out(out_file);
     }
